requirements.txt
torch>=2.10
torchvision>=0.25
pillow>=12
safetensors>=0.4
regex>=2024.0
jinja2>=3.1
pyarrow>=15
numpy>=1.26
wandb>=0.17  # optional, only needed when --wandb is passed
